# openapi.yaml
openapi: 3.0.3
info:
  title: Releasarr API
  version: 0.1.0
  description: |
    Contract derived from the new frontend client. Captures the HTTP interface the
    UI expects today so the backend can be validated and mocked consistently.
  license:
    name: Proprietary
    url: https://github.com/koreetz/releasarr
servers:
  - url: https://api.releasarr.dev/api
    description: Production server
  - url: http://localhost:8001/api
    description: Local development server
security:
  - BearerAuth: []
  - ServiceKeyAuth: []
paths:
  /requests:
    get:
      tags: [Requests]
      summary: List media requests
      operationId: listRequests
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/RequestStatus'
        - $ref: '#/components/parameters/RequestType'
        - $ref: '#/components/parameters/RequestOwner'
      responses:
        '200':
          description: Paged list of media requests.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestsResponse'
        '400':
          description: Invalid pagination or filter parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The caller is restricted to their own requests and may not filter by owner.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Requests]
      summary: Create a media request
      operationId: createRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaRequestCreate'
      responses:
        '201':
          description: Created request details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaRequest'
        '400':
          description: Malformed request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation failed for the provided fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /requests/{requestId}:
    get:
      tags: [Requests]
      summary: Retrieve a single request
      operationId: getRequest
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Requested media request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaRequest'
        '404':
          description: Request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Requests]
      summary: Partially update a media request
      operationId: updateRequest
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaRequestUpdate'
      responses:
        '200':
          description: Updated request details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaRequest'
        '400':
          description: Malformed request body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation failed for at least one field.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Requests]
      summary: Delete a media request
      operationId: deleteRequest
      description: |
        Removes the request and unmonitors what produced it: the season in Sonarr
        or the movie in Radarr. Dropping the row alone would not last, because the
        recurring sync rebuilds a request for every monitored season Sonarr still
        reports as missing.

        The series or movie itself stays in the library, as do any files already
        imported, and a series stays monitored even once its last season has
        gone - Sonarr reads one with no monitored season as wanting nothing.
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '204':
          description: Request removed.
        '404':
          description: Request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Sonarr or Radarr could not be reached.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /requests/{requestId}/releases:
    get:
      tags: [Releases]
      summary: List releases attached to a request
      operationId: listReleasesForRequest
      deprecated: true
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/ReleaseStatus'
      responses:
        '200':
          description: Releases linked to the provided request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleasesResponse'
        '400':
          description: Invalid pagination or filter parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /releases:
    get:
      tags: [Releases]
      summary: List releases
      operationId: listReleases
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/ReleaseStatus'
        - $ref: '#/components/parameters/RequestIdFilter'
      responses:
        '200':
          description: List of releases.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleasesResponse'
        '400':
          description: Invalid pagination or filter parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Releases]
      summary: Register a new release sourced from an external download
      operationId: createRelease
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddReleaseRequest'
      responses:
        '201':
          description: Created release metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Release'
        '400':
          description: Malformed request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: A release with the same identifier already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation failed for the provided fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /releases/{releaseId}:
    get:
      tags: [Releases]
      summary: Retrieve release details
      operationId: getRelease
      parameters:
        - $ref: '#/components/parameters/ReleaseId'
      responses:
        '200':
          description: Release metadata and files.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Release'
        '404':
          description: Release not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Releases]
      summary: Remove a release
      operationId: deleteRelease
      parameters:
        - $ref: '#/components/parameters/ReleaseId'
      responses:
        '204':
          description: Release removed.
        '404':
          description: Release not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /releases/{releaseId}/pause:
    post:
      tags: [Releases]
      summary: Pause an active release
      operationId: pauseRelease
      parameters:
        - $ref: '#/components/parameters/ReleaseId'
      responses:
        '202':
          description: Pause accepted.
          headers:
            Location:
              description: URL to poll for the async pause operation status.
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncOperationResponse'
        '400':
          description: Release cannot be paused because the request was invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Release not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Release is not in a pausable state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /releases/{releaseId}/resume:
    post:
      tags: [Releases]
      summary: Resume a paused release
      operationId: resumeRelease
      parameters:
        - $ref: '#/components/parameters/ReleaseId'
      responses:
        '202':
          description: Resume accepted.
          headers:
            Location:
              description: URL to poll for the async resume operation status.
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncOperationResponse'
        '400':
          description: Release cannot be resumed because the request was invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Release not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Release is not in a resumable state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /releases/search:
    get:
      tags: [Releases]
      summary: Search for release sources
      operationId: searchReleaseSources
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
          description: Free-text query string.
        - in: query
          name: request_id
          required: false
          schema:
            type: string
          description: Identifier of the media request initiating the search.
      responses:
        '200':
          description: Release source search results.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleaseSearchResponse'
        '400':
          description: Malformed search query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token obtained from /auth/login or /auth/refresh.
    ServiceKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        A service API key, acting as the user it is bound to. May be combined
        with X-Act-As-User to act as a different user, when the key allows it.
  parameters:
...
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
      description: Results page to retrieve (1-indexed).
    PerPage:
      in: query
      name: per_page
      schema:
        type: integer
        minimum: 1
        maximum: 100
      description: Number of items to return per page.
    RequestStatus:
      in: query
      name: status
      schema:
        $ref: '#/components/schemas/MediaRequestStatus'
      description: Filter requests by lifecycle status.
    RequestType:
      in: query
      name: type
      schema:
        $ref: '#/components/schemas/MediaType'
      description: Filter requests by media type.
    RequestOwner:
      in: query
      name: owner
      schema:
        type: string
      description: >-
        Filter requests by owner user id. Only callers with view_all_requests
        (or an admin) may use this; others are restricted to their own.
...